requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken
chainlit
python-dotenv
lxml
